# lib/api-zod/src/generated/types/autoCorrectFailure.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Career Pilot API
 * OpenAPI spec version: 0.1.0
 */
import type { AutoCorrectResult } from "./autoCorrectResult";

export interface AutoCorrectFailure {
  error: string;
  issues: string[];
  result?: AutoCorrectResult;
}